sender/capturer/requirements.txt
pyzmq
numpy
pyrealsense2
